+++ app/src/main/java/com/application/arenda/Entities/Announcements/Models/ModelAllAnnouncement.java
@@ -9,8 +9,8 @@ public class ModelAllAnnouncement {
     private int countRent = 0;
     private String placementDate;
 
-    private int idUser = 0;
-    private int idAnnouncement = 0;
+    private long idUser = 0;
+    private long idAnnouncement = 0;
 
     private boolean isFavorite = false;
 
@@ -30,11 +30,11 @@ public class ModelAllAnnouncement {
         this.mainUriBitmap = mainUriBitmap;
     }
 
-    public int getIdAnnouncement() {
+    public long getIdAnnouncement() {
         return idAnnouncement;
     }
 
-    public void setIdAnnouncement(int idAnnouncement) {
+    public void setIdAnnouncement(long idAnnouncement) {
         this.idAnnouncement = idAnnouncement;
     }
 
@@ -78,11 +78,11 @@ public class ModelAllAnnouncement {
         this.placementDate = placementDate;
     }
 
-    public int getIdUser() {
+    public long getIdUser() {
         return idUser;
     }
 
-    public void setIdUser(int idUser) {
+    public void setIdUser(long idUser) {
         this.idUser = idUser;
     }
