+++ app/src/main/java/com/application/arenda/Entities/Announcements/Models/ModelViewAnnouncement.java
@@ -5,8 +5,8 @@ import android.net.Uri;
 import java.util.List;
 
 public class ModelViewAnnouncement {
-    private int idUser = 0;
-    private int idAnnouncement = 0;
+    private long idUser = 0;
+    private long idAnnouncement = 0;
 
     private String name = "";
     private String description = "";
@@ -35,19 +35,19 @@ public class ModelViewAnnouncement {
 
     private List<Uri> uriCollection;
 
-    public int getIdUser() {
+    public long getIdUser() {
         return idUser;
     }
 
-    public void setIdUser(int idUser) {
+    public void setIdUser(long idUser) {
         this.idUser = idUser;
     }
 
-    public int getIdAnnouncement() {
+    public long getIdAnnouncement() {
         return idAnnouncement;
     }
 
-    public void setIdAnnouncement(int idAnnouncement) {
+    public void setIdAnnouncement(long idAnnouncement) {
         this.idAnnouncement = idAnnouncement;
     }
